--- src/io/teamelite/brush_menu/events/BrushItem.java
@@ -1,11 +1,0 @@
-package io.teamelite.brush_menu.events;
-
-/**
- * @name WorldMe
- * @authors Kieron Wiltshire
- * @contact kieron.wiltshire@outlook.com
- * @copyright Copyright (c) 2015 Kieron Wiltshire
- * @license MIT License
- */
-public class BrushItem {
-}
